requirements.txt
PyPDF2>=3.0.0
genanki>=0.13.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.8.0
httpx>=0.26.0
rich>=13.0.0
customtkinter>=5.2.0
pillow>=10.0.0
pdfplumber>=0.10.0
spacy>=3.7.0
pytest>=8.0.0
pyinstaller>=6.3.0 